+++ HISTORY.rst
@@ -30,6 +30,7 @@ latest (unreleased)
 * ALCYN-1844: Reservation not properly updated in case of stock increase and picking already partially served
 
 * ALCYN-1785: do not delay jobs for generating pdf files on supplier invoices
+* ALCYN-1847: Fix free product promotion on sales to be given only to allowed customer
 
 **Build**